requirements.txt
# FastAPI and web server
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0

# Vector database and embeddings - USING FAISS WITH SENTENCE TRANSFORMERS
faiss-cpu>=1.7.0,<1.8.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0

# Data validation and environment
pydantic>=2.0.0
python-dotenv

# HTTP requests for URL downloads - FIXED FOR RENDER DEPLOYMENT
httpx[http2]>=0.24.0

# Numerical operations and data handling
numpy>=1.24.0
scipy>=1.10.0

# Caching and performance
cachetools>=5.0.0
joblib>=1.3.0

# Enhanced logging and utilities
pathlib2>=2.3.7; python_version < '3.4'

# Optional: For better regex and text processing
regex>=2023.0.0

# Development and testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# For better error handling and monitoring (optional)
sentry-sdk[fastapi]>=1.32.0